+++ project_3/src/Item/RecordPlayer.java
@@ -1,7 +1,7 @@
 package Item;
 
 public class RecordPlayer extends Item{
-    public RecordPlayer (String itemName, int itemCondition, Float purchasePrice, Float listPrice, Float salePrice) {
+    public RecordPlayer (String itemName, int itemCondition, double purchasePrice, double listPrice, double salePrice) {
         super(itemName, itemCondition, purchasePrice, listPrice, salePrice);
     }
 }
